refactor path parser access
